services/incident-ai/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
httpx==0.28.1
nats-py==2.9.0
structlog==25.2.0
asyncpg==0.30.0
qdrant-client==1.13.3
pytest==8.3.5
pytest-asyncio==0.26.0
respx==0.22.0